requirements.txt
# ============================================================
# MULTI-AGENT CODE REVIEWER — Python Dependencies
# ============================================================
# Install: pip install -r requirements.txt

# --- Core: Anthropic API ---
anthropic>=0.40.0
python-dotenv>=1.0.0

# --- Multi-Agent Orchestration (LangGraph) ---
langgraph>=0.2.0
langchain-core>=0.3.0

# --- GitHub Integration ---
PyGithub>=2.3.0

# --- Utilities ---
requests>=2.31.0
